# coreService/coreapi/backend/swagger.json
{"swagger": "2.0", "info": {"title": "EduScroll Core", "description": "API documentation", "version": "v1"}, "host": "127.0.0.1:8000", "schemes": ["http"], "basePath": "/", "consumes": ["application/json"], "produces": ["application/json"], "securityDefinitions": {"Basic": {"type": "basic"}}, "security": [{"Basic": []}], "paths": {"/embed": {"get": {"operationId": "embed_list", "description": "", "parameters": [], "responses": {"200": {"description": ""}}, "tags": ["embed"]}, "post": {"operationId": "embed_create", "description": "", "parameters": [], "responses": {"201": {"description": ""}}, "tags": ["embed"]}, "parameters": []}, "/problem/{subject}": {"get": {"operationId": "problem_read", "description": "", "parameters": [], "responses": {"200": {"description": ""}}, "tags": ["problem"]}, "parameters": [{"name": "subject", "in": "path", "required": true, "type": "string"}]}, "/query_cache/add": {"get": {"operationId": "query_cache_add_list", "description": "", "parameters": [], "responses": {"200": {"description": ""}}, "tags": ["query_cache"]}, "post": {"operationId": "query_cache_add_create", "description": "", "parameters": [], "responses": {"201": {"description": ""}}, "tags": ["query_cache"]}, "parameters": []}, "/query_cache/get_reccomendation": {"get": {"operationId": "query_cache_get_reccomendation_list", "description": "", "parameters": [], "responses": {"200": {"description": ""}}, "tags": ["query_cache"]}, "post": {"operationId": "query_cache_get_reccomendation_create", "description": "", "parameters": [], "responses": {"201": {"description": ""}}, "tags": ["query_cache"]}, "parameters": []}, "/query_cache/search": {"get": {"operationId": "query_cache_search_list", "description": "", "parameters": [], "responses": {"200": {"description": ""}}, "tags": ["query_cache"]}, "post": {"operationId": "query_cache_search_create", "description": "", "parameters": [], "responses": {"201": {"description": ""}}, "tags": ["query_cache"]}, "parameters": []}, "/query_cache/status/add/{uuid}": {"get": {"operationId": "query_cache_status_add_read", "description": "", "parameters": [], "responses": {"200": {"description": ""}}, "tags": ["query_cache"]}, "parameters": [{"name": "uuid", "in": "path", "required": true, "type": "string"}]}, "/query_cache/status/get/{uuid}": {"get": {"operationId": "query_cache_status_get_read", "description": "", "parameters": [], "responses": {"200": {"description": ""}}, "tags": ["query_cache"]}, "parameters": [{"name": "uuid", "in": "path", "required": true, "type": "string"}]}, "/video": {"post": {"operationId": "video_create", "description": "", "parameters": [], "responses": {"201": {"description": ""}}, "tags": ["video"]}, "parameters": []}}, "definitions": {}}